Move messaging into internal package
