+++ internal/api/api.go
@@ -5,9 +5,9 @@ import (
 	"encoding/base64"
 	"encoding/json"
 	"github.com/creekorful/trandoshan/api"
+	"github.com/creekorful/trandoshan/internal/messaging"
 	"github.com/creekorful/trandoshan/internal/util/logging"
 	natsutil "github.com/creekorful/trandoshan/internal/util/nats"
-	"github.com/creekorful/trandoshan/messaging"
 	"github.com/labstack/echo/v4"
 	"github.com/nats-io/nats.go"
 	"github.com/olivere/elastic/v7"
