+++ internal/scheduler/scheduler.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"github.com/PuerkitoBio/purell"
 	"github.com/creekorful/trandoshan/api"
+	"github.com/creekorful/trandoshan/internal/messaging"
 	"github.com/creekorful/trandoshan/internal/util/logging"
 	natsutil "github.com/creekorful/trandoshan/internal/util/nats"
-	"github.com/creekorful/trandoshan/messaging"
 	"github.com/nats-io/nats.go"
 	"github.com/rs/zerolog/log"
 	"github.com/urfave/cli/v2"
